Update: Apply factory pattern on content/UnitTypes.java
